feat: Resolve 93 compilation errors through systematic ultrathinking approach
- Fixed missing struct field definitions and constructors across all modules
- Resolved type conversion errors and mismatches preserving data semantics
- Implemented missing methods with consciousness-aligned logic
- Fixed complex borrowing conflicts and ownership issues
- Corrected import paths and module organization
- Added complete enum variant implementations and pattern matching
- Preserved psychedelic consciousness simulation architecture:
  * Crystal consciousness system (Memory, Social, Quantum, Resonance, Chaos)
  * Species evolution mechanics (Disco Llamas, Quantum Sheep, Hypno Camels)
  * Mathematical chaos engines with 11D decision matrices
  * User co-evolution system with adaptive learning
  * Audio-visual synchronization through beat engines
  * Territory-based consciousness amplification

Library now compiles successfully with only warnings remaining.
Binary target requires additional import path fixes.

+++ src/mathematics/physics.rs
@@ -91,7 +91,7 @@ impl RealityField {
                 }
                 TearType::ChaosVortex => {
                     // Chaotic swirling motion
-                    tear.radius = 20.0 + (age * 3.0).sin() * 10.0;
+                    tear.radius = (20.0 + (age * 3.0).sin() * 10.0) as f32;
                     tear.strength = (((age * 7.0).cos().abs() * 0.3 + 0.7) *
                                   (1.0 - (age * 0.1).min(1.0))) as f32;
                 }
